refactor: move jw types to jw folder
