+++ src/types/index.d.ts
@@ -3,9 +3,9 @@ export * from './dates';
 export * from './electron';
 export * from './fs';
 export * from './general';
-export * from './lang';
+export * from './jw/lang';
+export * from './jw/publications';
+export * from './jw/sqlite';
 export * from './media';
 export * from './panzoom';
-export * from './publications';
 export * from './settings';
-export * from './sqlite';
